Allow item price to be decimal and not under 0

+++ app/src/Entity/Item.php
@@ -25,7 +25,7 @@ class Item
     private $title;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="decimal", precision=5, scale=2)
      */
     private $priceInfo;
 
@@ -68,12 +68,12 @@ class Item
         return $this;
     }
 
-    public function getPriceInfo(): ?int
+    public function getPriceInfo(): ?float
     {
         return $this->priceInfo;
     }
 
-    public function setPriceInfo(int $priceInfo): self
+    public function setPriceInfo(float $priceInfo): self
     {
         $this->priceInfo = $priceInfo;
